refactor: #159 materialPageable mapper extracted
